pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-core = "^0.3.6"
langchain-community = "^0.3.1"
langchain = "^0.3.1"
google = "^3.0.0"
genai = "^2.1.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
langchain-google-genai = "^2.0.0"
faiss-gpu = "^1.7.2"
faiss-cpu = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
